fixup: restructure config module to match conventions
