+++ app/src/config/constants.js
@@ -1,0 +1,14 @@
+// @flow
+
+import type { DevInternalFlag } from './types'
+
+export const DEV_INTERNAL_FLAGS: Array<DevInternalFlag> = [
+  'allPipetteConfig',
+  'enableMultiGEN2',
+  'enableBundleUpload',
+]
+
+// action type constants
+
+export const UPDATE: 'config:UPDATE' = 'config:UPDATE'
+export const RESET: 'config:RESET' = 'config:RESET'
